Se crea la funcionalidad de administración de opciones para un perfil. Se crean seeds para opciones y para opcionPerfil. Se adiciona el funcionamiento de las opciones de perfil al menu para que solo aparecan opciones permitidas por perfil. Se elimina la opción de registro en el login. Se eliminan del menu las opciones de editar, ver, etc.

+++ 03.Fuentes/resources/views/Opcion/index.blade.php
@@ -38,12 +38,17 @@ Listado Opciones Menu
             <td class="col-md-4">
                 <table>
                     <tr>
+                       @if (RayoCarroHelper::MostrarSubmenu(Auth::user()->perfil,'Opciones','Ver'))
                         <td>
                             <a href="{{ route('opcions.show', $opcion->id) }}" class="btn btn-primary">Ver</a>
                         </td>
+                        @endif
+                        @if (RayoCarroHelper::MostrarSubmenu(Auth::user()->perfil,'Opciones','Editar'))
                         <td>
                             <a href="{{ route('opcions.edit', $opcion->id) }}" class="btn btn-primary">Editar</a>
                         </td>
+                        @endif
+                        @if (RayoCarroHelper::MostrarSubmenu(Auth::user()->perfil,'Opciones','Eliminar'))
                         <td>
                             {!! Form::open([
                             'method' => 'DELETE',
@@ -52,6 +57,7 @@ Listado Opciones Menu
                             {!! Form::submit('Eliminar', ['class' => 'btn btn-danger']) !!}
                             {!! Form::close() !!}
                         </td>
+                        @endif
                     </tr>
                 </table>
             </td>
